#Mod 店铺权重 && 店铺权重系数

+++ src/main/java/com/tjlou/mybatis/auto/mysql/sps/entity/AuctionWeightShopRatioInfo.java
@@ -14,7 +14,7 @@ import java.io.Serializable;
  * </p>
  *
  * @author Gaby
- * @since 2020-11-03
+ * @since 2020-12-23
  */
 @TableName("auction_weight_shop_ratio_info")
 public class AuctionWeightShopRatioInfo implements Serializable, IEntity {
@@ -26,9 +26,7 @@ public class AuctionWeightShopRatioInfo implements Serializable, IEntity {
     @TableId(value = "id", type = IdType.AUTO)
     private Long id;
     /**
-     * 系数类型  1-店铺首页被分享的人数  2-店铺商品被分享的人数 3-通过永久二维码访问的人数 4-通过商品二维码访问的人数  5-店铺首页分享链接访问的人数
-     * 6-商品分享链接访问的人数  7-店铺付款成交额
-     * 8-店铺付款成交人数  9-店铺付款成交比数 10-店铺当面交易成交额 11-店铺退款额  12-店铺退款额人数
+     * 系数类型  1-店铺首页被分享的人数  2-店铺商品被分享的人数 3-通过永久二维码访问的人数 4-通过商品二维码访问的人数  5-店铺首页分享链接访问的人数  6-商品分享链接访问的人数  7-店铺付款成交额   8-店铺付款成交人数  9-店铺付款成交比数 10-店铺当面交易成交额 11-店铺退款额  12-店铺退款额人数
      */
     @TableField("ratio_type")
     private Integer ratioType;
@@ -36,7 +34,7 @@ public class AuctionWeightShopRatioInfo implements Serializable, IEntity {
      * 系数值
      */
     @TableField("ratio_value")
-    private Integer ratioValue;
+    private Double ratioValue;
     /**
      * 系数天数之内
      */
@@ -62,11 +60,11 @@ public class AuctionWeightShopRatioInfo implements Serializable, IEntity {
         return this;
     }
 
-    public Integer getRatioValue() {
+    public Double getRatioValue() {
         return ratioValue;
     }
 
-    public AuctionWeightShopRatioInfo setRatioValue(Integer ratioValue) {
+    public AuctionWeightShopRatioInfo setRatioValue(Double ratioValue) {
         this.ratioValue = ratioValue;
         return this;
     }
@@ -90,9 +88,7 @@ public class AuctionWeightShopRatioInfo implements Serializable, IEntity {
     public static final String ID = "id";
 
     /**
-     * 系数类型  1-店铺首页被分享的人数  2-店铺商品被分享的人数 3-通过永久二维码访问的人数 4-通过商品二维码访问的人数  5-店铺首页分享链接访问的人数
-     * 6-商品分享链接访问的人数  7-店铺付款成交额
-     * 8-店铺付款成交人数  9-店铺付款成交比数 10-店铺当面交易成交额 11-店铺退款额  12-店铺退款额人数
+     * 系数类型  1-店铺首页被分享的人数  2-店铺商品被分享的人数 3-通过永久二维码访问的人数 4-通过商品二维码访问的人数  5-店铺首页分享链接访问的人数  6-商品分享链接访问的人数  7-店铺付款成交额   8-店铺付款成交人数  9-店铺付款成交比数 10-店铺当面交易成交额 11-店铺退款额  12-店铺退款额人数
      */
     public static final String RATIO_TYPE = "ratio_type";
